Add get_traits() method to ComputeDriver
The get_traits() virt driver method can be used by virt driver
implementations to advertise the traits that are associated with the
resource provider representing a given compute node. If implemented, the
method should return an iterable of traits.

Adds a call to the ComputeDriver.get_traits() method in the resource
tracker, and submits the traits to the placement API using the scheduler
client. Not all compute drivers implement get_traits, so the resource
tracker gracefully handles NotImplementedError.

blueprint ironic-driver-traits

Change-Id: Ia088341d86aa87f72c9decb081e0a43c36939f18