Unplug all VIFs from ironic nodes during tear down
In some rare cases nova can lose track of some of an instances VIFs,
preventing them from being detached from the ironic node. This can
cause subsequent deployments to the node to fail with this error:

VirtualInterfacePlugException: Cannot attach VIF
409830a5-b4de-4d1d-be22-5e6fe4ccd65b to the node
3aaaf79e-99fb-42a3-b22e-b1a7fae44272 due to error: Unable to attach VIF
409830a5-b4de-4d1d-be22-5e6fe4ccd65b, not enough free physical ports.

To work around this, when tearing down an instance we now list all VIFs
for the node in ironic after the known instance VIFs have been detached,
and detach those as well.

Change-Id: I70077f5f249ee35bfd4a08a9180172f571b1cb5a
Closes-Bug: #1733861
Related-Bug: #1743652