Fix ironic instance rebuild with API >=2.93
When you attempt to rebuild an Ironic instance, using the CLI,
it will pick the latest micversion, which triggers sending
to the ironic driver:
reimage_boot_volume = true

The ironic driver then raises an exception, saying it is unable
to rebuild a volume backed instance. This happens even when
the instance is not volume backed, and rebuild would be fine.

This patch ensures we only raise the "can't rebuild"
exception when the instance is volume backed.

Closes-Bug: #2127017

Change-Id: Ibb8e83a1d36506df34a324ca0da9e2f129a38cfa
Signed-off-by: John Garbutt <john.garbutt@stackhpc.com>